# tickets-back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Biblionacho API",
        "description": "Listado de URI'S para la API de Biblionacho",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/tickets/index": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "All tickets",
                "description": "Show all  tickets",
                "operationId": "IndexTickets",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "Bearer id|YOUR_ACCESS_TOKEN_HERE"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ok 200"
                                        },
                                        "Tickets": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Ticket"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server Error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "errors": [
                                                    "Internal server error!"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "NotFound"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "errors": [
                                                    "Tickets not found."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tickets/store": {
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Create a ticket",
                "description": "Create a ticket",
                "operationId": "StoreTicket",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "Bearer\n     *   id|YOUR_ACCESS\n     *   TOKEN_HERE\n     *  "
                        }
                    }
                ],
                "requestBody": {
                    "description": " of ticket",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title ",
                                    "description",
                                    "deadline",
                                    "user_id",
                                    "status_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "implement CRUD for ticket management"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Don't forget to implement a deadline for the ticket"
                                    },
                                    "deadline": {
                                        "type": "string",
                                        "example": "24-05-2024"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "status_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully registered"
                                        },
                                        "response": {
                                            "$ref": "#/components/schemas/Ticket"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "title": [
                                                    "The title field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server Error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "errors": [
                                                    "Internal server error!"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Ticket": {
                "required": [
                    "title ",
                    "description",
                    "deadline",
                    "user_id",
                    "status_id"
                ],
                "properties": {
                    "title": {
                        "description": "title of ticket",
                        "type": "string",
                        "example": "implement CRUD for ticket management"
                    },
                    "description": {
                        "description": "Don't forget to implement a deadline for the ticket",
                        "type": "string",
                        "example": ""
                    },
                    "deadline": {
                        "description": "deadline of the  ticket",
                        "type": "string",
                        "example": "24-05-2024"
                    },
                    "user_id": {
                        "description": "User responsible for resolving the ticket",
                        "type": "integer",
                        "example": 1
                    },
                    "status_id": {
                        "description": "status ID of the  ticket",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            }
        }
    }
}